TOY-55 PreviewActivity 리팩토링
