+++ app/src/main/res/layout/activity_preview.xml
@@ -1,85 +1,94 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".camera.capture.PreviewActivity">
+    xmlns:tools="http://schemas.android.com/tools">
 
-    <android.support.v7.widget.Toolbar
-        android:id="@+id/toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="?attr/actionBarSize"
-        android:background="@android:color/transparent"
-        app:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar">
+    <data>
 
-        <ImageButton
-            android:id="@+id/btn_back_capture"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:background="?selectableItemBackgroundBorderless"
-            android:src="@drawable/toolbar_back" />
-    </android.support.v7.widget.Toolbar>
-
-    <ImageView
-        android:id="@+id/preview_image"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_weight="1"
-        android:adjustViewBounds="true"
-        android:visibility="visible" />
+        <variable
+            name="viewModel"
+            type="kr.co.yogiyo.rookiephotoapp.camera.capture.PreviewViewModel" />
+    </data>
 
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_marginTop="10dp"
-        android:layout_marginBottom="20dp"
-        android:layout_weight="2.5"
-        android:orientation="horizontal">
+        android:orientation="vertical"
+        tools:context=".camera.capture.PreviewActivity">
 
-        <Button
-            android:id="@+id/btn_add_diary"
+        <android.support.v7.widget.Toolbar
+            android:id="@+id/toolbar"
+            android:layout_width="match_parent"
+            android:layout_height="?attr/actionBarSize"
+            android:background="@android:color/transparent"
+            app:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar">
+
+            <ImageButton
+                android:id="@+id/btn_back_capture"
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:background="?selectableItemBackgroundBorderless"
+                android:src="@drawable/toolbar_back" />
+        </android.support.v7.widget.Toolbar>
+
+        <ImageView
+            android:id="@+id/preview_image"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:layout_marginLeft="30dp"
-            android:layout_marginTop="30dp"
-            android:layout_marginBottom="20dp"
-            android:layout_weight="2"
-            android:background="?selectableItemBackgroundBorderless"
-            android:drawableTop="@mipmap/diary_add_gray_"
-            android:drawablePadding="-5dp"
-            android:text="글쓰기"
-            android:textColor="#8F9699"
-            android:textStyle="bold" />
-
-        <Button
-            android:id="@+id/btn_save_photo"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_marginLeft="70dp"
-            android:layout_marginTop="20dp"
-            android:layout_marginRight="70dp"
-            android:layout_marginBottom="20dp"
             android:layout_weight="1"
-            android:background="@mipmap/photo_save" />
+            android:adjustViewBounds="true" />
 
-        <Button
-            android:id="@+id/btn_edit"
+        <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:layout_marginTop="30dp"
-            android:layout_marginRight="30dp"
+            android:layout_marginTop="10dp"
             android:layout_marginBottom="20dp"
-            android:layout_weight="2"
-            android:background="?selectableItemBackgroundBorderless"
-            android:drawableTop="@mipmap/photo_edit"
-            android:drawablePadding="-5dp"
-            android:text="@string/text_edit"
-            android:textColor="#8F9699"
-            android:textStyle="bold" />
+            android:layout_weight="2.5"
+            android:orientation="horizontal">
+
+            <Button
+                android:id="@+id/btn_add_diary"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:layout_marginLeft="30dp"
+                android:layout_marginTop="30dp"
+                android:layout_marginBottom="20dp"
+                android:layout_weight="2"
+                android:background="?selectableItemBackgroundBorderless"
+                android:drawableTop="@mipmap/diary_add_gray_"
+                android:drawablePadding="-5dp"
+                android:text="@string/text_write_diary"
+                android:textColor="#8F9699"
+                android:textStyle="bold" />
+
+            <Button
+                android:id="@+id/btn_save_photo"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:layout_marginLeft="70dp"
+                android:layout_marginTop="20dp"
+                android:layout_marginRight="70dp"
+                android:layout_marginBottom="20dp"
+                android:layout_weight="1"
+                android:background="@mipmap/photo_save" />
+
+            <Button
+                android:id="@+id/btn_edit"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:layout_marginTop="30dp"
+                android:layout_marginRight="30dp"
+                android:layout_marginBottom="20dp"
+                android:layout_weight="2"
+                android:background="?selectableItemBackgroundBorderless"
+                android:drawableTop="@mipmap/photo_edit"
+                android:drawablePadding="-5dp"
+                android:text="@string/text_edit"
+                android:textColor="#8F9699"
+                android:textStyle="bold" />
+
+        </LinearLayout>
+
 
     </LinearLayout>
-
-
-</LinearLayout>
+</layout>
