+++ app/src/main/res/values/strings.xml
@@ -101,4 +101,7 @@
     <string name="text_day_now_diary">%d일</string>
     <string name="text_time_now_diary">%d:%s</string>
     <string name="text_yogidiary_path">%s/%s</string>
+
+    <!--activity_preview.xml-->
+    <string name="text_write_diary">글쓰기</string>
 </resources>
